requirements.txt
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
Flask==1.0.2
Flask-PyMongo==2.2.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pillow==6.0.0
pycparser==2.19
pymongo==3.7.2
pytesseract==0.2.6
six==1.12.0
tesseract==0.1.3
Werkzeug==0.15.1
